fixed comments 28/11
