+++ src/main/java/com/solvd/carina/demo/mobile/gui/pages/common/LoginSwagPageBase.java
@@ -10,8 +10,6 @@ public abstract class LoginSwagPageBase extends AbstractPage {
 
     public abstract WelcomePageBase login(String username, String password);
 
-    public abstract void loginRandomCredentials(String randomUsername, String randomPass);
-
     public abstract boolean isLoginErrorPromptPresent();
 
     public abstract boolean isPassWordInputPresent();
